fix: update prepare_for_reboot to handle tracing
 - a new status code WAITING_ON_REBOOT
 - properly using the volume_api to remove volmes rather than assuming
   docker volumes.
 - add a test for prepare_to_reboot to make sure it works going forward

+++ jobrunner/cli/prepare_for_reboot.py
@@ -4,32 +4,38 @@ automatically re-run after a reboot.
 """
 import argparse
 
-from jobrunner.executors.local import container_name, docker
-from jobrunner.executors.volumes import docker_volume_name as volume_name
-from jobrunner.lib.database import find_where, update_where
-from jobrunner.models import Job, State
+from jobrunner.executors.local import container_name, docker, volume_api
+from jobrunner.lib.database import find_where
+from jobrunner.models import Job, State, StatusCode
+from jobrunner.run import set_state
 
 
-def main():
-    print(
-        "== DANGER ZONE ==\n"
-        "\n"
-        "This will kill all running jobs and reset them to the PENDING state, ready\n"
-        "to be restarted following a reboot.\n"
-        "\n"
-        "It should only be run when the job-runner service has been stopped."
-        "\n"
-    )
-    confirm = input("Are you sure you want to continue? (y/N)")
-    assert confirm.strip().lower() == "y"
-    # Reset all running jobs to pending
-    update_where(Job, {"state": State.PENDING, "started_at": None}, state=State.RUNNING)
-    # Make sure all containers and volumes are removed ready to freshly restart the jobs
-    # after the reboot
-    for job in find_where(Job, state=State.PENDING):
+def main(pause=True):
+    if pause:
+        print(
+            "== DANGER ZONE ==\n"
+            "\n"
+            "This will kill all running jobs and reset them to the PENDING state, ready\n"
+            "to be restarted following a reboot.\n"
+            "\n"
+            "It should only be run when the job-runner service has been stopped."
+            "\n"
+        )
+        confirm = input("Are you sure you want to continue? (y/N)")
+        assert confirm.strip().lower() == "y"
+
+    for job in find_where(Job, state=State.RUNNING):
+        print(f"reseting job {job.id} to PENDING")
+        set_state(
+            job,
+            State.PENDING,
+            StatusCode.WAITING_ON_REBOOT,
+            "Job restarted - waiting for server to reboot",
+        )
+        # these are idempotent
         docker.kill(container_name(job))
         docker.delete_container(container_name(job))
-        docker.delete_volume(volume_name(job))
+        volume_api.delete_volume(job)
 
 
 def run():
